Update for Nav Bar
There are certain issues with the Nav bar that require us to confirm its position every time an update is made. For instance, in the previous version I worked on, the NavBar was fine, but after Zhilin updated the version, the Nav Bar became misplaced.

Currently, this version has been written by Heitung, and I am responsible for updating it within this branch.